@@ -7,21 +7,25 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@expo/vector-icons": "^10.0.6",
-    "expo": "^36.0.0",
-    "expo-asset": "~8.0.0",
-    "expo-linear-gradient": "~8.0.0",
-    "react": "16.9.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.1.tar.gz",
-    "react-native-gesture-handler": "^1.5.3",
-    "react-native-screens": "^2.0.0-beta.7",
+    "@expo/vector-icons": "^12.0.0",
+    "expo": "^40.0.0",
+    "expo-app-loading": "^1.0.1",
+    "expo-asset": "~8.2.1",
+    "expo-linear-gradient": "~8.4.0",
+    "native-base": "^2.13.15",
+    "react": "16.13.1",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.0.tar.gz",
+    "react-native-gesture-handler": "~1.8.0",
+    "react-native-gifted-chat": "^0.16.3",
+    "react-native-screens": "~2.15.0",
     "react-native-slider": "^0.11.0",
     "react-native-vector-icons": "^6.6.0",
     "react-navigation": "^4.0.10",
-    "react-navigation-stack": "^1.9.3"
+    "react-navigation-stack": "^1.9.3",
+    "socket.io-client": "^3.0.4"
   },
   "devDependencies": {
-    "babel-preset-expo": "^8.0.0"
+    "babel-preset-expo": "8.3.0"
   },
   "private": true
 }
