@@ -1,16 +1,17 @@
-import React from 'react';
-import {View, Text} from 'react-native';
-import { GiftedChat } from 'react-native-gifted-chat'
+import React, {Component , useCallback,}from 'react';
+//import { View, KeyboardAvoidingView, TextInput, StyleSheet, Text, Platform, TouchableWithoutFeedback, Button, Keyboard  } from 'react-native';
+import { AppRegistry, View, StatusBar } from "react-native";
+import { Container, Body, Content, Header, Left, Right, Icon, Title, Input, Item, Label, Button, Text } from "native-base";
 import io from 'socket.io-client';
 const socket = io();
 
 
-export default class Codes extends Component {
+export default class Chat extends Component {
     
     state = {
-        messages: [],
+        sms:"",
         id:"",
-        id_group:""
+        id_group:"",
     }; 
     
     componentDidMount() {
@@ -35,41 +36,76 @@ export default class Codes extends Component {
             });  
         });
     }
-    async onSend(sms){
+    async onSend(){
         // const id = await AsyncStorage.getItem('id');
         // const id_group= await AsyncStorage.getItem('id_group');
-        const { id, id_group } = this.state;
+        const { id, id_group,sms } = this.state;
+        console.log(sms);
 
         var json ={
             "sms":sms,
             "id_user":id,
             "id_group":id_group
         }
-        socket.emit('mensaje', json);
+        await socket.emit('mensaje', json);
     }
+
+   
+    
     render () {
+        const { id_group } = this.state;
+        socket.on('mensaje/'+id_group, function (msj) { 
+            console.log(msj)
+            //$('#listado-msjs').append( $('<li>').text(msj) );
+        });
         return (
-            <GiftedChat
-            messages={messages}
-            onSend={(message) => {onSend(messages)}}
-            user={{
-              _id: 1,
-            }}
-          />
+            <Container>
+                <Header>
+                <Left>
+                    <Button
+                    transparent
+                    onPress={() => this.props.navigation.navigate("DrawerOpen")}>
+                    <Icon name="menu" />
+                    </Button>
+                </Left>
+                <Body>
+                    <Title>Lucy Chat</Title>
+                </Body>
+                <Right />
+                </Header>
+                <Content padder>
+                <Item floatingLabel style={{ marginTop: 300 }}>
+                    <Input 
+                    sms
+                    placeholder='Enter Message'
+                    onChangeText={msj => this.setState({ sms: msj })}
+                    />
+                    <Icon name='paper-plane' onPress={() => this.onSend()} />
+    
+                    {/* <Button rounded 
+                    // style={{ marginTop: 20, alignSelf: "center" }}
+                    // onPress={() => navigate("Profile")}
+                    >
+                        <Text>Send</Text>
+                    </Button> */}
+                </Item>
+                
+                </Content>
+            </Container>
         );
     }
 }
 
-chat.defaultProps = {
-    name: "fanny"
-};
+// chat.defaultProps = {
+//     name: "fanny"
+// };
 
-chat.propTypes = {
-    name: React.PropTypes.string,
-}
+// chat.propTypes = {
+//     name: React.PropTypes.string,
+// }
 
-const { id_group } = this.state;
-socket.on('mensaje/'+id_group, function (msj) { 
-    console.log(msj)
-    //$('#listado-msjs').append( $('<li>').text(msj) );
-});
+// const { id_group } = this.state;
+// socket.on('mensaje/'+id_group, function (msj) { 
+//     console.log(msj)
+//     //$('#listado-msjs').append( $('<li>').text(msj) );
+// });
