@@ -1,4 +1,4 @@
 /* export const ip= "https://try-2-tele-heroku.herokuapp.com/"
  */
 
-export const ip= "http://192.168.250.11:3000"
+export const ip= "http://192.168.250.13:3000"
