@@ -11,6 +11,7 @@ import loginCode from "../screens/loginCode";
 import loginNumber from "../screens/loginNumber";
 import Chat from '../screens/Chat'
 import Home from '../screens/Home'
+import Group from '../screens/Group'
 
 import { theme } from "../constants";
 
@@ -24,7 +25,8 @@ const screens = createStackNavigator(
     loginCode,
     Settings,
     Chat,
-    Home
+    Home,
+    Group
   },
   {
     defaultNavigationOptions: {
